Переносит генерацию данных и служебные функции в отдельные модули
